Add log listener and processor
``otorderd`` runs all the intensive tasks requested by the web app in
the background, with a multiprocessing.Queue listening on port 60000 for
requests from the web app.

``otorderd_logd`` is a true daemon process that logs stuff from
``otorderd``. A log daemon is necessary because there are multiple
processes logging to the same file, and it keeps the messages from being
interleaved.