Dataset sorting, saving error
No options to sort datasets in account view, but changed sorting to be
by upload date, with newest first. "Error saving unknown" came from
deleting constraints that were being used in a priori ranking. Easy fix
was to clear the ranking whenever a constraint is deleted. Not elegant
or perhaps intuitive, but will suffice for now.